fix runtime issues from incorrect type (closes #2)
The hid_t return value of H5Fopen was getting stuffed into an int, which
led to runtime errors whenever this handle is accessed. Not sure if this
only occurs with 1.10, or on stretch, or whatever, but the code was
clearly incorrect. The workaround is to not use the int member of the
pcmio struct and just open/close the file whenever it's accessed.